Add check for ConfigMgr client and run client actions
I realize this script is really meant for switching to Intune/Autopatch but thought including relevant ConfigMgr client actions would be useful since it would be writing to the same registry locations that GPO would and therefore could impact the outcome of this reset process. Feel free to discard if you don't feel it should be within scope.